Website Drafts - centralize project drafts
- Centralize the check for project drafts.
- Use that check when resolving document draft links.
- Update nav resolution to use it as well.
